changed the way that the enigma smart contract addresses are collected
Previously, when running in the discovery-cli environment, the ethereum
addresses of the Enigma and Enigma Token contracts were just collected
from the JSON files of their ABI, and when running in the new
docker-environment, they were collected by fetching from an http server
that publishes them inside that network.

Now, the addresses are expected to be found in environment variables on
the server running the operator, and the client in turn queries the
operator for their values.