Added the option to set the operator's eth account from the environment.
The operator codebase no longer assumes that it has access to
unlocked accounts, but rather tries to set a default account for
the web3 connection when initializing it. This account is either
generated from a private key passed in as the OPERATOR_ETH_PRIVATE_KEY
environment variable, or fetched from the list of unlocked accounts on
the ethereum node providing the connection to the ethereum network.
If no address could be generated using the described method, an error
is thrown.